+++ packages/camera/camera_android_camerax/lib/src/camerax_library.g.dart
@@ -124,6 +124,7 @@ class PigeonOverrides {
     ResolutionSelector? resolutionSelector,
     int? targetRotation,
     CameraXFlashMode? flashMode,
+    int? jpegQuality,
   })?
   imageCapture_new;
 
@@ -4078,12 +4079,14 @@ class ImageCapture extends UseCase {
     ResolutionSelector? resolutionSelector,
     int? targetRotation,
     CameraXFlashMode? flashMode,
+    int? jpegQuality,
   }) {
     if (PigeonOverrides.imageCapture_new != null) {
       return PigeonOverrides.imageCapture_new!(
         resolutionSelector: resolutionSelector,
         targetRotation: targetRotation,
         flashMode: flashMode,
+        jpegQuality: jpegQuality,
       );
     }
     return ImageCapture.pigeon_new(
@@ -4092,6 +4095,7 @@ class ImageCapture extends UseCase {
       resolutionSelector: resolutionSelector,
       targetRotation: targetRotation,
       flashMode: flashMode,
+      jpegQuality: jpegQuality,
     );
   }
 
@@ -4102,6 +4106,7 @@ class ImageCapture extends UseCase {
     this.resolutionSelector,
     int? targetRotation,
     CameraXFlashMode? flashMode,
+    int? jpegQuality,
   }) : super.pigeon_detached() {
     final int pigeonVar_instanceIdentifier = pigeon_instanceManager.addDartCreatedInstance(this);
     final _PigeonInternalProxyApiBaseCodec pigeonChannelCodec = _pigeonVar_codecImageCapture;
@@ -4118,6 +4123,7 @@ class ImageCapture extends UseCase {
       resolutionSelector,
       targetRotation,
       flashMode,
+      jpegQuality,
     ]);
     () async {
       final pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;