+++ packages/camera/camera_avfoundation/test/avfoundation_camera_test.mocks.dart
@@ -342,4 +342,13 @@ class MockCameraApi extends _i1.Mock implements _i2.CameraApi {
             returnValueForMissingStub: _i4.Future<void>.value(),
           )
           as _i4.Future<void>);
+
+  @override
+  _i4.Future<void> setJpegImageQuality(int? quality) =>
+      (super.noSuchMethod(
+            Invocation.method(#setJpegImageQuality, [quality]),
+            returnValue: _i4.Future<void>.value(),
+            returnValueForMissingStub: _i4.Future<void>.value(),
+          )
+          as _i4.Future<void>);
 }