+++ packages/google_sign_in/google_sign_in_android/pigeons/messages.dart
@@ -88,6 +88,7 @@ class PlatformGoogleIdTokenCredential {
   String? familyName;
   String? givenName;
   late String email;
+  late String uniqueId;
   late String idToken;
   String? profilePictureUri;
 }