+++ packages/pigeon/platform_tests/test_plugin/android/src/main/kotlin/com/example/test_plugin/CoreTests.gen.kt
@@ -251,6 +251,10 @@ data class UnusedClass(val aField: Any? = null) {
     result = 31 * result + CoreTestsPigeonUtils.deepHash(this.aField)
     return result
   }
+
+  override fun toString(): String {
+    return "UnusedClass(aField=$aField)"
+  }
 }
 
 /**
@@ -452,6 +456,10 @@ data class AllTypes(
     result = 31 * result + CoreTestsPigeonUtils.deepHash(this.listMap)
     result = 31 * result + CoreTestsPigeonUtils.deepHash(this.mapMap)
     return result
+  }
+
+  override fun toString(): String {
+    return "AllTypes(aBool=$aBool, anInt=$anInt, anInt64=$anInt64, aDouble=$aDouble, aByteArray=${aByteArray.contentToString()}, a4ByteArray=${a4ByteArray.contentToString()}, a8ByteArray=${a8ByteArray.contentToString()}, aFloatArray=${aFloatArray.contentToString()}, anEnum=$anEnum, anotherEnum=$anotherEnum, aString=$aString, anObject=$anObject, list=$list, stringList=$stringList, intList=$intList, doubleList=$doubleList, boolList=$boolList, enumList=$enumList, objectList=$objectList, listList=$listList, mapList=$mapList, map=$map, stringMap=$stringMap, intMap=$intMap, enumMap=$enumMap, objectMap=$objectMap, listMap=$listMap, mapMap=$mapMap)"
   }
 }
 
@@ -673,6 +681,10 @@ data class AllNullableTypes(
     result = 31 * result + CoreTestsPigeonUtils.deepHash(this.recursiveClassMap)
     return result
   }
+
+  override fun toString(): String {
+    return "AllNullableTypes(aNullableBool=$aNullableBool, aNullableInt=$aNullableInt, aNullableInt64=$aNullableInt64, aNullableDouble=$aNullableDouble, aNullableByteArray=${aNullableByteArray?.contentToString()}, aNullable4ByteArray=${aNullable4ByteArray?.contentToString()}, aNullable8ByteArray=${aNullable8ByteArray?.contentToString()}, aNullableFloatArray=${aNullableFloatArray?.contentToString()}, aNullableEnum=$aNullableEnum, anotherNullableEnum=$anotherNullableEnum, aNullableString=$aNullableString, aNullableObject=$aNullableObject, allNullableTypes=$allNullableTypes, list=$list, stringList=$stringList, intList=$intList, doubleList=$doubleList, boolList=$boolList, enumList=$enumList, objectList=$objectList, listList=$listList, mapList=$mapList, recursiveClassList=$recursiveClassList, map=$map, stringMap=$stringMap, intMap=$intMap, enumMap=$enumMap, objectMap=$objectMap, listMap=$listMap, mapMap=$mapMap, recursiveClassMap=$recursiveClassMap)"
+  }
 }
 
 /**
@@ -876,6 +888,46 @@ data class AllNullableTypesWithoutRecursion(
     result = 31 * result + CoreTestsPigeonUtils.deepHash(this.mapMap)
     return result
   }
+
+  override fun toString(): String {
+    return "AllNullableTypesWithoutRecursion(aNullableBool=$aNullableBool, aNullableInt=$aNullableInt, aNullableInt64=$aNullableInt64, aNullableDouble=$aNullableDouble, aNullableByteArray=${aNullableByteArray?.contentToString()}, aNullable4ByteArray=${aNullable4ByteArray?.contentToString()}, aNullable8ByteArray=${aNullable8ByteArray?.contentToString()}, aNullableFloatArray=${aNullableFloatArray?.contentToString()}, aNullableEnum=$aNullableEnum, anotherNullableEnum=$anotherNullableEnum, aNullableString=$aNullableString, aNullableObject=$aNullableObject, list=$list, stringList=$stringList, intList=$intList, doubleList=$doubleList, boolList=$boolList, enumList=$enumList, objectList=$objectList, listList=$listList, mapList=$mapList, map=$map, stringMap=$stringMap, intMap=$intMap, enumMap=$enumMap, objectMap=$objectMap, listMap=$listMap, mapMap=$mapMap)"
+  }
+}
+
+/**
+ * A data class without fields for testing empty classes.
+ *
+ * Generated class from Pigeon that represents data sent in messages.
+ */
+class AnEmptyClass {
+  companion object {
+    fun fromList(pigeonVar_list: List<Any?>): AnEmptyClass {
+      return AnEmptyClass()
+    }
+  }
+
+  fun toList(): List<Any?> {
+    return listOf()
+  }
+
+  override fun equals(other: Any?): Boolean {
+    if (other == null || other.javaClass != javaClass) {
+      return false
+    }
+    if (this === other) {
+      return true
+    }
+    return true
+  }
+
+  override fun hashCode(): Int {
+    var result = javaClass.hashCode()
+    return result
+  }
+
+  override fun toString(): String {
+    return "AnEmptyClass()"
+  }
 }
 
 /**
@@ -894,7 +946,8 @@ data class AllClassesWrapper(
     val classList: List<AllTypes?>,
     val nullableClassList: List<AllNullableTypesWithoutRecursion?>? = null,
     val classMap: Map<Long?, AllTypes?>,
-    val nullableClassMap: Map<Long?, AllNullableTypesWithoutRecursion?>? = null
+    val nullableClassMap: Map<Long?, AllNullableTypesWithoutRecursion?>? = null,
+    val anEmptyClass: AnEmptyClass? = null
 ) {
   companion object {
     fun fromList(pigeonVar_list: List<Any?>): AllClassesWrapper {
@@ -905,6 +958,7 @@ data class AllClassesWrapper(
       val nullableClassList = pigeonVar_list[4] as List<AllNullableTypesWithoutRecursion?>?
       val classMap = pigeonVar_list[5] as Map<Long?, AllTypes?>
       val nullableClassMap = pigeonVar_list[6] as Map<Long?, AllNullableTypesWithoutRecursion?>?
+      val anEmptyClass = pigeonVar_list[7] as AnEmptyClass?
       return AllClassesWrapper(
           allNullableTypes,
           allNullableTypesWithoutRecursion,
@@ -912,7 +966,8 @@ data class AllClassesWrapper(
           classList,
           nullableClassList,
           classMap,
-          nullableClassMap)
+          nullableClassMap,
+          anEmptyClass)
     }
   }
 
@@ -925,6 +980,7 @@ data class AllClassesWrapper(
         nullableClassList,
         classMap,
         nullableClassMap,
+        anEmptyClass,
     )
   }
 
@@ -943,7 +999,8 @@ data class AllClassesWrapper(
         CoreTestsPigeonUtils.deepEquals(this.classList, other.classList) &&
         CoreTestsPigeonUtils.deepEquals(this.nullableClassList, other.nullableClassList) &&
         CoreTestsPigeonUtils.deepEquals(this.classMap, other.classMap) &&
-        CoreTestsPigeonUtils.deepEquals(this.nullableClassMap, other.nullableClassMap)
+        CoreTestsPigeonUtils.deepEquals(this.nullableClassMap, other.nullableClassMap) &&
+        CoreTestsPigeonUtils.deepEquals(this.anEmptyClass, other.anEmptyClass)
   }
 
   override fun hashCode(): Int {
@@ -955,7 +1012,12 @@ data class AllClassesWrapper(
     result = 31 * result + CoreTestsPigeonUtils.deepHash(this.nullableClassList)
     result = 31 * result + CoreTestsPigeonUtils.deepHash(this.classMap)
     result = 31 * result + CoreTestsPigeonUtils.deepHash(this.nullableClassMap)
+    result = 31 * result + CoreTestsPigeonUtils.deepHash(this.anEmptyClass)
     return result
+  }
+
+  override fun toString(): String {
+    return "AllClassesWrapper(allNullableTypes=$allNullableTypes, allNullableTypesWithoutRecursion=$allNullableTypesWithoutRecursion, allTypes=$allTypes, classList=$classList, nullableClassList=$nullableClassList, classMap=$classMap, nullableClassMap=$nullableClassMap, anEmptyClass=$anEmptyClass)"
   }
 }
 
@@ -994,6 +1056,10 @@ data class TestMessage(val testList: List<Any?>? = null) {
     result = 31 * result + CoreTestsPigeonUtils.deepHash(this.testList)
     return result
   }
+
+  override fun toString(): String {
+    return "TestMessage(testList=$testList)"
+  }
 }
 
 private open class CoreTestsPigeonCodec : StandardMessageCodec() {
@@ -1020,9 +1086,12 @@ private open class CoreTestsPigeonCodec : StandardMessageCodec() {
         }
       }
       135.toByte() -> {
-        return (readValue(buffer) as? List<Any?>)?.let { AllClassesWrapper.fromList(it) }
+        return (readValue(buffer) as? List<Any?>)?.let { AnEmptyClass.fromList(it) }
       }
       136.toByte() -> {
+        return (readValue(buffer) as? List<Any?>)?.let { AllClassesWrapper.fromList(it) }
+      }
+      137.toByte() -> {
         return (readValue(buffer) as? List<Any?>)?.let { TestMessage.fromList(it) }
       }
       else -> super.readValueOfType(type, buffer)
@@ -1055,12 +1124,16 @@ private open class CoreTestsPigeonCodec : StandardMessageCodec() {
         stream.write(134)
         writeValue(stream, value.toList())
       }
-      is AllClassesWrapper -> {
+      is AnEmptyClass -> {
         stream.write(135)
         writeValue(stream, value.toList())
       }
-      is TestMessage -> {
+      is AllClassesWrapper -> {
         stream.write(136)
+        writeValue(stream, value.toList())
+      }
+      is TestMessage -> {
+        stream.write(137)
         writeValue(stream, value.toList())
       }
       else -> super.writeValue(stream, value)
@@ -1099,6 +1172,14 @@ interface HostIntegrationCoreApi {
   fun echoObject(anObject: Any): Any
   /** Returns the passed list, to test serialization and deserialization. */
   fun echoList(list: List<Any?>): List<Any?>
+  /** Returns the passed list, to test serialization and deserialization. */
+  fun echoStringList(stringList: List<String?>): List<String?>
+  /** Returns the passed list, to test serialization and deserialization. */
+  fun echoIntList(intList: List<Long?>): List<Long?>
+  /** Returns the passed list, to test serialization and deserialization. */
+  fun echoDoubleList(doubleList: List<Double?>): List<Double?>
+  /** Returns the passed list, to test serialization and deserialization. */
+  fun echoBoolList(boolList: List<Boolean?>): List<Boolean?>
   /** Returns the passed list, to test serialization and deserialization. */
   fun echoEnumList(enumList: List<AnEnum?>): List<AnEnum?>
   /** Returns the passed list, to test serialization and deserialization. */
@@ -1808,6 +1889,94 @@ interface HostIntegrationCoreApi {
             val wrapped: List<Any?> =
                 try {
                   listOf(api.echoList(listArg))
+                } catch (exception: Throwable) {
+                  CoreTestsPigeonUtils.wrapError(exception)
+                }
+            reply.reply(wrapped)
+          }
+        } else {
+          channel.setMessageHandler(null)
+        }
+      }
+      run {
+        val channel =
+            BasicMessageChannel<Any?>(
+                binaryMessenger,
+                "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi.echoStringList$separatedMessageChannelSuffix",
+                codec)
+        if (api != null) {
+          channel.setMessageHandler { message, reply ->
+            val args = message as List<Any?>
+            val stringListArg = args[0] as List<String?>
+            val wrapped: List<Any?> =
+                try {
+                  listOf(api.echoStringList(stringListArg))
+                } catch (exception: Throwable) {
+                  CoreTestsPigeonUtils.wrapError(exception)
+                }
+            reply.reply(wrapped)
+          }
+        } else {
+          channel.setMessageHandler(null)
+        }
+      }
+      run {
+        val channel =
+            BasicMessageChannel<Any?>(
+                binaryMessenger,
+                "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi.echoIntList$separatedMessageChannelSuffix",
+                codec)
+        if (api != null) {
+          channel.setMessageHandler { message, reply ->
+            val args = message as List<Any?>
+            val intListArg = args[0] as List<Long?>
+            val wrapped: List<Any?> =
+                try {
+                  listOf(api.echoIntList(intListArg))
+                } catch (exception: Throwable) {
+                  CoreTestsPigeonUtils.wrapError(exception)
+                }
+            reply.reply(wrapped)
+          }
+        } else {
+          channel.setMessageHandler(null)
+        }
+      }
+      run {
+        val channel =
+            BasicMessageChannel<Any?>(
+                binaryMessenger,
+                "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi.echoDoubleList$separatedMessageChannelSuffix",
+                codec)
+        if (api != null) {
+          channel.setMessageHandler { message, reply ->
+            val args = message as List<Any?>
+            val doubleListArg = args[0] as List<Double?>
+            val wrapped: List<Any?> =
+                try {
+                  listOf(api.echoDoubleList(doubleListArg))
+                } catch (exception: Throwable) {
+                  CoreTestsPigeonUtils.wrapError(exception)
+                }
+            reply.reply(wrapped)
+          }
+        } else {
+          channel.setMessageHandler(null)
+        }
+      }
+      run {
+        val channel =
+            BasicMessageChannel<Any?>(
+                binaryMessenger,
+                "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi.echoBoolList$separatedMessageChannelSuffix",
+                codec)
+        if (api != null) {
+          channel.setMessageHandler { message, reply ->
+            val args = message as List<Any?>
+            val boolListArg = args[0] as List<Boolean?>
+            val wrapped: List<Any?> =
+                try {
+                  listOf(api.echoBoolList(boolListArg))
                 } catch (exception: Throwable) {
                   CoreTestsPigeonUtils.wrapError(exception)
                 }